+++ coresdk/src/backend/input_driver.h
@@ -58,10 +58,10 @@ namespace splashkit_lib
     sk_window_data sk_get_window_event_data(sk_drawing_surface *surface);
     void sk_move_window(sk_drawing_surface *surface, int x, int y);
 
-    void sk_start_reading_text(window wind, float x, float y, float width, float height, string initial_text);
+    void sk_start_reading_text(window wind, double x, double y, double width, double height, string initial_text);
 
-    void sk_mouse_position(float &x, float &y);
-    void sk_mouse_movement(float &x, float &y);
+    void sk_mouse_position(double &x, double &y);
+    void sk_mouse_movement(double &x, double &y);
     
     bool sk_mouse_button_down(uint32_t button);
