+++ coresdk/src/coresdk/circle_drawing.cpp
@@ -22,7 +22,7 @@
 namespace splashkit_lib
 {
 
-    void draw_circle(color clr, float x, float y, float radius, drawing_options opts)
+    void draw_circle(color clr, double x, double y, double radius, drawing_options opts)
     {
         sk_drawing_surface *surface;
 
@@ -36,7 +36,7 @@ namespace splashkit_lib
         }
     }
 
-    void draw_circle(color clr, float x, float y, float radius)
+    void draw_circle(color clr, double x, double y, double radius)
     {
         draw_circle(clr, x, y, radius, option_defaults());
     }
@@ -51,7 +51,7 @@ namespace splashkit_lib
         draw_circle(clr, c.center.x, c.center.y, c.radius, option_defaults());
     }
 
-    void fill_circle(color clr, float x, float y, float radius, drawing_options opts)
+    void fill_circle(color clr, double x, double y, double radius, drawing_options opts)
     {
         sk_drawing_surface *surface;
 
@@ -63,7 +63,7 @@ namespace splashkit_lib
             sk_fill_circle(surface, clr, x, y, abs(radius));
     }
 
-    void fill_circle(color clr, float x, float y, float radius)
+    void fill_circle(color clr, double x, double y, double radius)
     {
         fill_circle(clr, x, y, radius, option_defaults());
     }
@@ -78,24 +78,23 @@ namespace splashkit_lib
         fill_circle(clr, c.center.x, c.center.y, c.radius, option_defaults());
     }
     
-    void draw_circle_on_window(window destination, color clr, float x, float y, float radius, drawing_options opts)
+    void draw_circle_on_window(window destination, color clr, double x, double y, double radius, drawing_options opts)
     {
         draw_circle(clr, x, y, radius, option_draw_to(destination, opts));
     }
     
-    void draw_circle_on_window(window destination, color clr, float x, float y, float radius)
+    void draw_circle_on_window(window destination, color clr, double x, double y, double radius)
     {
         draw_circle(clr, x, y, radius, option_draw_to(destination));
     }
     
-    void draw_circle_on_bitmap(bitmap destination, color clr, float x, float y, float radius, drawing_options opts)
+    void draw_circle_on_bitmap(bitmap destination, color clr, double x, double y, double radius, drawing_options opts)
     {
         draw_circle(clr, x, y, radius, option_draw_to(destination, opts));
     }
     
-    void draw_circle_on_bitmap(bitmap destination, color clr, float x, float y, float radius)
+    void draw_circle_on_bitmap(bitmap destination, color clr, double x, double y, double radius)
     {
         draw_circle(clr, x, y, radius, option_draw_to(destination));
     }
-    
 }
